Athena Connection Methods

Implementations of pure virtual functions defined in the DBI package for AthenaConnection objects.

Method to get Athena schema, tables and table types return as a data.frame

This method returns all partitions from Athena table.

Executes a statement to return the data description language (DDL) of the Athena table.

# S4 method for class 'AthenaConnection'
show(object)

# S4 method for class 'AthenaConnection'
dbDisconnect(conn, ...)

# S4 method for class 'AthenaConnection'
dbIsValid(dbObj, ...)

# S4 method for class 'AthenaConnection,character'
dbSendQuery(conn, statement, unload = athena_unload(), ...)

# S4 method for class 'AthenaConnection,character'
dbSendStatement(conn, statement, unload = athena_unload(), ...)

# S4 method for class 'AthenaConnection,character'
dbExecute(conn, statement, unload = athena_unload(), ...)

# S4 method for class 'AthenaConnection,ANY'
dbDataType(dbObj, obj, ...)

# S4 method for class 'AthenaConnection,data.frame'
dbDataType(dbObj, obj, ...)

# S4 method for class 'AthenaConnection,character'
dbQuoteString(conn, x, ...)

# S4 method for class 'AthenaConnection,POSIXct'
dbQuoteString(conn, x, ...)

# S4 method for class 'AthenaConnection,Date'
dbQuoteString(conn, x, ...)

# S4 method for class 'AthenaConnection,SQL'
dbQuoteIdentifier(conn, x, ...)

dbGetTables(conn, ...)

# S4 method for class 'AthenaConnection'
dbGetTables(conn, catalog = NULL, schema = NULL, ...)

# S4 method for class 'AthenaConnection,character'
dbListFields(conn, name, ...)

# S4 method for class 'AthenaConnection,character'
dbExistsTable(conn, name, ...)

# S4 method for class 'AthenaConnection,Id'
dbExistsTable(conn, name, ...)

# S4 method for class 'AthenaConnection,character'
dbRemoveTable(conn, name, delete_data = TRUE, confirm = FALSE, ...)

# S4 method for class 'AthenaConnection,Id'
dbRemoveTable(conn, name, delete_data = TRUE, confirm = FALSE, ...)

# S4 method for class 'AthenaConnection,character'
dbGetQuery(conn, statement, statistics = FALSE, unload = athena_unload(), ...)

# S4 method for class 'AthenaConnection'
dbGetInfo(dbObj, ...)

dbGetPartition(conn, name, ..., .format = FALSE)

# S4 method for class 'AthenaConnection'
dbGetPartition(conn, name, ..., .format = FALSE)

dbShow(conn, name, ...)

# S4 method for class 'AthenaConnection'
dbShow(conn, name, ...)

# S4 method for class 'AthenaConnection'
dbBegin(conn, ...)

# S4 method for class 'AthenaConnection'
dbCommit(conn, ...)

# S4 method for class 'AthenaConnection'
dbRollback(conn, ...)

Arguments

object

Any R object

conn

A DBI::DBIConnection object, as returned by dbConnect().

...

Other parameters passed on to methods.

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult.

statement

a character string containing SQL.

unload

boolean input to modify statement to align with AWS Athena UNLOAD, default is set to FALSE.

obj

An R object whose SQL type we want to determine.

x

A character vector to quote as string.

catalog

Athena catalog, default set to NULL to return all tables from all Athena catalogs

schema

Athena schema, default set to NULL to return all tables from all Athena schemas. Note: The use of DATABASE and SCHEMA is interchangeable within Athena.

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

delete_data

Deletes S3 files linking to AWS Athena table

confirm

Allows for S3 files to be deleted without the prompt check. It is recommend to leave this set to FALSE to avoid deleting other S3 files when the table's definition points to the root of S3 bucket.

statistics

If set to TRUE will print out AWS Athena statistics of query.

.format

re-formats AWS Athena partitions format. So that each column represents a partition from the AWS Athena table. Default set to FALSE to prevent breaking previous package behaviour.

Value

dbGetTables() returns a data.frame.

data.frame that returns all partitions in table, if no partitions in Athena table then function will return error from Athena.

dbShow() returns SQL characters of the Athena table DDL.

Slots

ptr

a list of connecting objects from the SDK paws package.

info

a list of metadata objects

quote

syntax to quote sql query when creating Athena ddl