S3 implementation of db_desc for Athena (api version 1).

This is a backend function for dplyr to retrieve meta data about Athena queries. Users won't be required to access and run this function.

db_desc.AthenaConnection(x)

Arguments

x

A dbConnect object, as returned by dbConnect()

Value

Character variable containing Meta Data about query sent to Athena. The Meta Data is returned in the following format:

"Athena <paws version> [<profile_name>@region/database]"