Creates query to create a simple Athena table

Creates an interface to compose CREATE EXTERNAL TABLE.

# S4 method for AthenaConnection
sqlCreateTable(con, table = NULL,
  fields = NULL, field.types = NULL, partition = NULL,
  s3.location = NULL, file.type = c("csv", "tsv", "parquet"), ...)

Arguments

con
A database connection.

table
Name of the table. Escaped with dbQuoteIdentifier().

fields
Either a character vector or a data frame.

A named character vector: Names are column names, values are types. Names are escaped with dbQuoteIdentifier(). Field types are unescaped.

A data frame: field types are generated using dbDataType().

field.types
Additional field types used to override derived types.

partition
Partition Athena table (needs to be a named list or vector) for example: c(var1 = "2019-20-13")

s3.location
s3 bucket to store Athena table

file.type
What file type to store data.frame on s3, noctua currently supports ["csv", "tsv", "parquet"]

...
Other arguments used by individual methods.

Value

sqlCreateTable returns data.frame's DDL in the SQL format.

See also

Examples

# \donttest{
# Note: 
# - Require AWS Account to run below example.
# - Different connection methods can be used please see `noctua::dbConnect` documnentation

library(DBI)

# Demo connection to Athena using profile name 
con <- dbConnect(noctua::athena())

# Create DDL for iris data.frame
sqlCreateTable(con, "iris", iris, s3.location = "s3://path/to/athena/table")

#> Info: data.frame names have been converted to align with Athena DDL naming convertions: 
#> sepal_length,
#> sepal_width,
#> petal_length,
#> petal_width,
#> species

#> <SQL> CREATE EXTERNAL TABLE iris (
#>   sepal_length DOUBLE,
#>   sepal_width DOUBLE,
#>   petal_length DOUBLE,
#>   petal_width DOUBLE,
#>   species STRING
#> )
#> ROW FORMAT DELIMITED
#> 	FIELDS TERMINATED BY ','
#> 	ESCAPED BY '\\'
#> 	LINES TERMINATED BY '\n'
#> LOCATION 's3://path/to/athena/table/iris/'
#> TBLPROPERTIES ("skip.header.line.count"="1");

# Create DDL for iris data.frame with partition
sqlCreateTable(con, "iris", iris,
               partition = c("timestamp" = format(Sys.Date(), "%Y%m%d")),
               s3.location = "s3://path/to/athena/table")

#> Info: data.frame names have been converted to align with Athena DDL naming convertions: 
#> sepal_length,
#> sepal_width,
#> petal_length,
#> petal_width,
#> species

#> <SQL> CREATE EXTERNAL TABLE iris (
#>   sepal_length DOUBLE,
#>   sepal_width DOUBLE,
#>   petal_length DOUBLE,
#>   petal_width DOUBLE,
#>   species STRING
#> )
#> PARTITIONED BY (20191016 STRING)
#> ROW FORMAT DELIMITED
#> 	FIELDS TERMINATED BY ','
#> 	ESCAPED BY '\\'
#> 	LINES TERMINATED BY '\n'
#> LOCATION 's3://path/to/athena/table/iris/'
#> TBLPROPERTIES ("skip.header.line.count"="1");

# Create DDL for iris data.frame with partition and file.type parquet
sqlCreateTable(con, "iris", iris,
               partition = c("timestamp" = format(Sys.Date(), "%Y%m%d")),
               s3.location = "s3://path/to/athena/table",
               file.type = "parquet")

#> Info: data.frame names have been converted to align with Athena DDL naming convertions: 
#> sepal_length,
#> sepal_width,
#> petal_length,
#> petal_width,
#> species

#> <SQL> CREATE EXTERNAL TABLE iris (
#>   sepal_length DOUBLE,
#>   sepal_width DOUBLE,
#>   petal_length DOUBLE,
#>   petal_width DOUBLE,
#>   species STRING
#> )
#> PARTITIONED BY (20191016 STRING)
#> STORED AS PARQUET
#> LOCATION 's3://path/to/athena/table/iris/'
#> ;

# Disconnect from Athena
dbDisconnect(con)
# }